move hovercard.js
